Add StacktraceBuilder for building stacktrace
This moves the construction logic away from StacktraceInterface and
Event classes. It allows users to customize stacktrace construction easier.
