+++ sentry-ruby/lib/sentry/interfaces/stacktrace.rb
@@ -2,17 +2,8 @@ module Sentry
   class StacktraceInterface
     attr_reader :frames
 
-    def initialize(backtrace:, project_root:, app_dirs_pattern:, linecache:, context_lines:, backtrace_cleanup_callback: nil)
-      @frames = []
-
-      parsed_backtrace_lines = Backtrace.parse(
-        backtrace, project_root, app_dirs_pattern, &backtrace_cleanup_callback
-      ).lines
-
-      parsed_backtrace_lines.reverse.each_with_object(@frames) do |line, frames|
-        frame = convert_parsed_line_into_frame(line, project_root, linecache, context_lines)
-        frames << frame if frame.filename
-      end
+    def initialize(frames)
+      @frames = frames
     end
 
     def to_hash
@@ -20,12 +11,6 @@ module Sentry
     end
 
     private
-
-    def convert_parsed_line_into_frame(line, project_root, linecache, context_lines)
-      frame = StacktraceInterface::Frame.new(project_root, line)
-      frame.set_context(linecache, context_lines) if context_lines
-      frame
-    end
 
     # Not actually an interface, but I want to use the same style
     class Frame < Interface
